+++ lib/roda/project/jeremy_template_creator.rb
@@ -1,4 +1,4 @@
-require 'fileutils'
+require "fileutils"
 
 class Roda
   module Project
@@ -12,7 +12,7 @@ class Roda
         success = system("git clone --depth 1 #{repo_url} #{@context.project_name}")
         abort("\nCould not download the template.") unless success
 
-        git_dir = File.join(@context.project_name, '.git')
+        git_dir = File.join(@context.project_name, ".git")
         if Dir.exist?(git_dir)
           FileUtils.rm_rf(git_dir)
           puts "\n* Cleaned up template git history.\n\n"
@@ -24,10 +24,9 @@ class Roda
         system(setup_command)
 
         action_success_message(setup_command, "run")
-        puts "\n Setup is done! follow the link below to learn about this template:"
-        puts "\n #{repo}"
+        puts "\nSetup is done! follow the link below to learn about this template:"
+        puts "\n#{repo_url}"
       end
     end
   end
 end
-